+++ src/app/api-doc/page.tsx
@@ -1,10 +1,10 @@
-import { getApiDocs } from '@/lib/swagger';
 import ReactSwagger from './react-swagger';
+import { getApiDocs } from '@/lib/swagger';
 
 export default async function ApiDocPage() {
   const spec = await getApiDocs();
   return (
-    <section className="w-full h-screen">
+    <section className="w-full">
       <ReactSwagger spec={spec} />
     </section>
   );
