+++ src/app/api-doc/react-swagger.tsx
@@ -1,7 +1,8 @@
 'use client';
 
-import SwaggerUI from 'swagger-ui-react';
 import 'swagger-ui-react/swagger-ui.css';
+
+import SwaggerUI from 'swagger-ui-react';
 
 type Props = {
   spec: Record<string, any>;
