+++ src/app/api/register/route.ts
@@ -1,5 +1,5 @@
-import bcrypt from 'bcrypt';
 import { NextResponse } from 'next/server';
+import bcrypt from 'bcrypt';
 import { prisma } from '@/lib/prisma';
 
 export async function POST(request: Request) {
