+++ src/app/dashboard/page.tsx
@@ -1,9 +1,9 @@
 'use client';
 
-import { useSession } from 'next-auth/react';
-import { useRouter } from 'next/navigation';
-import { useState } from 'react';
 import { Button } from '@/components/general/button';
+import { useRouter } from 'next/navigation';
+import { useSession } from 'next-auth/react';
+import { useState } from 'react';
 
 interface PostData {
   title: string;
