+++ src/components/general/icon/icons/index.ts
@@ -1,13 +1,13 @@
+import GithubIcon from './github-icon';
 import MoonFilledIcon from './moon-filled-icon';
+import SearchIcon from './search-icon';
 import SunFilledIcon from './sun-filled-icon';
 import XIcon from './x-icon';
-import GithubIcon from './github-icon';
-import SearchIcon from './search-icon';
 
 export const iconNameMap = {
-  MoonFilledIcon,
-  SunFilledIcon,
-  XIcon,
   GithubIcon,
-  SearchIcon
+  MoonFilledIcon,
+  SearchIcon,
+  SunFilledIcon,
+  XIcon
 } as const;
